requirements.txt
aiogram==3.5.0
openai==1.30.5
python-dotenv==1.1.1
